refactor(hickey): relocate loopback out of links/ — it's not a link
links/loopback.ts self-documented 'not a link' while sitting in links/ and
exporting as @kolu/surface/links/loopback — the directory asserted a category
it disclaims. Move to src/loopback.ts, export @kolu/surface/loopback. (Lowy
cross-validation corrected the destination: a one-member src/transport/ dir
would be a speculative axis; top-level beside the family is right.)
